Grading
This is pretty sloppy,

Code you state as broken is still around, at the very least comment it out, I had to make quite a few edits to get your Rmd to Knit. And even then it took forever because your choice of lines map was HUGE. In the future if you're going to turn in something this large please knit it before sending it to me. I hit you pretty hard on functionality and code

Your first map is fine, but the second one is minimal and takes forever to render...